Friday Backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "friday-assistant"
version = "1.0.0"
description = "FRIDAY Voice Assistant - Voice-controlled AI assistant with system control capabilities"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Railey", email = "you@example.com"}
]
requires-python = ">=3.8"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "websockets>=11.0",
    "python-dotenv>=1.0.0",
    "google-genai>=0.1.0",
    "pyttsx3>=2.90",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
vision = [
    "opencv-python>=4.8",
    "mediapipe>=0.10",
    "pyautogui>=0.9",
]

[project.urls]
Homepage = "https://github.com/Railey29/Friday"
Repository = "https://github.com/Railey29/Friday"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
